Add fullscreen mode for mermaid
