+++ components/content/Mermaid.vue
@@ -15,53 +15,24 @@
   -->
 
 <script setup lang="ts">
-const { $mermaid } = useNuxtApp()
+import FullscreenView from '../ep/FullscreenView.vue'
+import MermaidDiagram from '../ep/MermaidDiagram.vue'
 
 const props = defineProps<{
   code: string
 }>()
-
-const decodedCode = computed(() => {
-  return atob(props.code)
-})
-
-const root = ref<HTMLElement | null>(null)
-const codeBlock = ref<HTMLElement | null>(null)
-const isDiagramLoading = ref(true)
-
-async function renderMermaidDiagram() {
-  isDiagramLoading.value = true
-  if (codeBlock.value && $mermaid) {
-    try {
-      await $mermaid.run({
-        nodes: [codeBlock.value],
-        suppressErrors: true
-      })
-    } catch (e) {}
-
-    isDiagramLoading.value = false
-  }
-}
-
-onMounted(() => {
-  renderMermaidDiagram()
-})
 </script>
 
 <template>
-  <figure ref="root" class="relative">
-    <pre
-      ref="codeBlock"
-      style="background: none"
-      :class="{
-        'opacity-0': isDiagramLoading
-      }"
-      v-text="decodedCode"
-    ></pre>
-    <div>
-      <div v-if="isDiagramLoading" class="absolute inset-0 font-serif">
-        Mermaid diagram is loading...
-      </div>
-    </div>
-  </figure>
+  <FullscreenView>
+    <template #default="{ onClick, isFullscreen }">
+      <MermaidDiagram
+        :class="{
+          'cursor-zoom-in hover:shadow transition-shadow': !isFullscreen
+        }"
+        :code="props.code"
+        @click="onClick"
+      />
+    </template>
+  </FullscreenView>
 </template>
